feat: :sparkles: added extra fns to namespaces
