+++ test/scripts/bspNetBootstrap.ts
@@ -56,7 +56,7 @@ async function bootStrapNetwork() {
 
   api = await BspNetTestApi.create(`ws://127.0.0.1:${NODE_INFOS.user.port}`);
 
-  await api.sendNewStorageRequest(CONFIG.localPath, CONFIG.remotePath, CONFIG.bucketName);
+  await api.file.newStorageRequest(CONFIG.localPath, CONFIG.remotePath, CONFIG.bucketName);
 
   await api.wait.bspVolunteer();
   await api.wait.bspStored();
