+++ test/util/asserts.ts
@@ -3,35 +3,36 @@ import { strictEqual } from "node:assert";
 import type { ApiPromise } from "@polkadot/api";
 import type { AugmentedEvent } from "@polkadot/api/types";
 
+export type AssertExtrinsicOptions = {
+  /** The block height to check. If not provided, the latest block will be used. */
+  blockHeight?: string;
+  /** The block hash to check. Takes precedence over blockHeight if provided. */
+  blockHash?: string;
+  /** If true, skips the check for an associated `ExtrinsicSuccess` event. */
+  skipSuccessCheck?: boolean;
+  /** If true, checks the pending transaction pool instead of a finalized block. */
+  checkTxPool?: boolean;
+  /** The module name of the extrinsic to check (e.g., "balances"). */
+  module: string;
+  /** The method name of the extrinsic to check (e.g., "transfer"). */
+  method: string;
+  /** If true, skips the validation check for the module.method existence in the API metadata. */
+  ignoreParamCheck?: boolean;
+};
+
 /**
  * Asserts that a specific extrinsic (module.method) is present in a blockchain block or transaction pool.
  *
- * @param {ApiPromise} api - The API instance connected to the blockchain network.
- * @param {Object} options - Configuration options for the extrinsic check.
- * @param {string} [options.blockHeight] - The block height to check. If not provided, the latest block will be used.
- * @param {string} [options.blockHash] - The block hash to check. Takes precedence over blockHeight if provided.
- * @param {boolean} [options.skipSuccessCheck=false] - If true, skips the check for an associated `ExtrinsicSuccess` event.
- * @param {boolean} [options.checkTxPool=false] - If true, checks the pending transaction pool instead of a finalized block.
- * @param {string} options.module - The module name of the extrinsic to check (e.g., "balances").
- * @param {string} options.method - The method name of the extrinsic to check (e.g., "transfer").
- * @param {boolean} [options.ignoreParamCheck=false] - If true, skips the validation check for the module.method existence in the API metadata.
- *
- * @returns {Promise<Object[]>} - Returns a list of objects representing the extrinsics that match the module.method criteria.
- * @throws {Error} - Throws an error if no matching extrinsic is found, or if the success check fails (unless skipped).
+ * @param api The API instance connected to the blockchain network.
+ * @param options Configuration options for the extrinsic check.
+ * @returns A list of objects representing the extrinsics that match the module.method criteria.
+ * @throws Error if no matching extrinsic is found, or if the success check fails (unless skipped).
  *
  * TODO: add ability to search nested extrinsics e.g. sudo.sudo(balance.forceTransfer(...))
  */
 export const assertExtrinsicPresent = async (
   api: ApiPromise,
-  options: {
-    blockHeight?: string;
-    blockHash?: string;
-    skipSuccessCheck?: boolean;
-    checkTxPool?: boolean;
-    module: string;
-    method: string;
-    ignoreParamCheck?: boolean;
-  }
+  options: AssertExtrinsicOptions
 ): Promise<
   {
     module: string;
@@ -95,13 +96,12 @@ export const assertExtrinsicPresent = async (
 /**
  * Asserts that a specific event (module.method) is present in the provided list of events.
  *
- * @param {ApiPromise} api - The API instance connected to the blockchain network.
- * @param {string} module - The module name of the event to check (e.g., "system").
- * @param {string} method - The method name of the event to check (e.g., "ExtrinsicSuccess").
- * @param {EventRecord[]} [events] - The list of events to search through. If not provided or empty, an error is thrown.
- *
- * @returns {Object} - Returns an object containing the matching event and its data.
- * @throws {Error} - Throws an error if no matching event is found, or if the event does not match the expected structure.
+ * @param api The API instance connected to the blockchain network.
+ * @param module The module name of the event to check (e.g., "system").
+ * @param method The method name of the event to check (e.g., "ExtrinsicSuccess").
+ * @param events The list of events to search through. If not provided or empty, an error is thrown.
+ * @returns An object containing the matching event and its data.
+ * @throws Error if no matching event is found, or if the event does not match the expected structure.
  */
 export const assertEventPresent = (
   api: ApiPromise,
@@ -130,13 +130,12 @@ export const assertEventPresent = (
 /**
  * Asserts that multiple instances of a specific event (module.method) are present in the provided list of events.
  *
- * @param {ApiPromise} api - The API instance connected to the blockchain network.
- * @param {string} module - The module name of the event to check (e.g., "system").
- * @param {string} method - The method name of the event to check (e.g., "ExtrinsicSuccess").
- * @param {EventRecord[]} [events] - The list of events to search through. If not provided or empty, an error is thrown.
- *
- * @returns {EventRecord[]} - Returns an array of matching events.
- * @throws {Error} - Throws an error if no matching events are found.
+ * @param api The API instance connected to the blockchain network.
+ * @param module The module name of the event to check (e.g., "system").
+ * @param method The method name of the event to check (e.g., "ExtrinsicSuccess").
+ * @param events The list of events to search through. If not provided or empty, an error is thrown.
+ * @returns An array of matching events.
+ * @throws Error if no matching events are found.
  */
 export const assertEventMany = (
   api: ApiPromise,
@@ -190,4 +189,5 @@ export namespace Assertions {
   export const eventPresent = assertEventPresent;
   export const eventMany = assertEventMany;
   export const fetchEvent = fetchEventData;
+  export const extrinsicPresent = assertExtrinsicPresent;
 }
