+++ test/util/bspNet/index.ts
@@ -7,3 +7,4 @@ export * from "./testrunner";
 export * from "./toxic";
 export * from "./waits";
 export * from "./test-api";
+export * from "./fileHelpers";
