+++ test/util/bspNet/test-api.ts
@@ -5,8 +5,15 @@ import type { EventRecord } from "@polkadot/types/interfaces";
 import type { ISubmittableResult } from "@polkadot/types/types";
 import "@storagehub/api-augment";
 import { types as BundledTypes } from "@storagehub/types-bundle";
-import { assertEventMany, assertEventPresent, Assertions } from "../asserts";
-import { createBucket, sealBlock, sendNewStorageRequest } from "./helpers";
+import {
+  assertEventMany,
+  assertEventPresent,
+  assertExtrinsicPresent,
+  Assertions,
+  type AssertExtrinsicOptions
+} from "../asserts";
+import { createBucket, Files, sendNewStorageRequest } from "./fileHelpers";
+import { sealBlock } from "./helpers";
 import type { BspNetApi } from "./types";
 import { waitForBspStored, waitForBspVolunteer, Waits } from "./waits";
 
@@ -57,7 +64,9 @@ export class BspNetTestApi implements AsyncDisposable {
       eventPresent: (module: string, method: string, events?: EventRecord[]) =>
         assertEventPresent(this._api, module, method, events),
       eventMany: (module: string, method: string, events?: EventRecord[]) =>
-        assertEventMany(this._api, module, method, events)
+        assertEventMany(this._api, module, method, events),
+      extrinsicPresent: (options: AssertExtrinsicOptions) =>
+        assertExtrinsicPresent(this._api, options)
     };
 
     const remappedWaitsNs = {
@@ -66,22 +75,42 @@ export class BspNetTestApi implements AsyncDisposable {
       bspStored: () => waitForBspStored(this._api)
     };
 
+    const remappedFileNs = {
+      ...Files,
+      /**
+       * Creates a new bucket.
+       *
+       * @param bucketName - The name of the bucket to be created.
+       * @returns A promise that resolves to a new bucket event.
+       */
+      newBucket: (bucketName: string) => createBucket(this._api, bucketName),
+      /**
+       * Creates a new bucket and submits a new storage request.
+       *
+       * @param source - The local path to the file to be uploaded.
+       * @param location - The StorageHub "location" field of the file to be uploaded.
+       * @param bucketName - The name of the bucket to be created.
+       * @returns A promise that resolves to file metadata.
+       */
+      newStorageRequest: (source: string, location: string, bucketName: string) =>
+        sendNewStorageRequest(this._api, source, location, bucketName)
+    };
+
     return Object.assign(this._api, {
       sealBlock: this.sealBlock.bind(this),
       sendNewStorageRequest: this.sendNewStorageRequest.bind(this),
       createBucket: this.createBucket.bind(this),
       assertEvent: this.assertEvent.bind(this),
       assert: remappedAssertNs,
-      wait: remappedWaitsNs
+      wait: remappedWaitsNs,
+      file: remappedFileNs
+      // Add docker
     }) satisfies BspNetApi;
   }
 
   async [Symbol.asyncDispose]() {
     await this._api.disconnect();
   }
-
-  // TODO: Add namespaces
-  //      - files
 }
 
 export type EnrichedBspApi = Awaited<ReturnType<typeof BspNetTestApi.create>>;
